Add graph link to ration
